suggested revisions
