+++ Project.toml
@@ -26,7 +26,7 @@ ClimaParams = "0.10"
 Cloudy = "0.4"
 DataFrames = "1.6"
 DocStringExtensions = "0.8, 0.9"
-EnsembleKalmanProcesses = "1.1.5"
+EnsembleKalmanProcesses = "1.1"
 ForwardDiff = "0.10"
 MLJ = "0.20"
 RootSolvers = "0.3, 0.4"
