+++ src/AerosolActivation.jl
@@ -23,8 +23,7 @@ export mean_hygroscopicity_parameter,
     N_activated_per_mode,
     M_activated_per_mode,
     total_N_activated,
-    total_M_activated,
-    CalibratedAerosolActivationParameters
+    total_M_activated
 
 """
     coeff_of_curvature(ap, T)
@@ -312,13 +311,5 @@ function total_M_activated(
 ) where {FT}
     return sum(M_activated_per_mode(ap, ad, aip, tps, T, p, w, q))
 end
-
-
-"""
-    CalibratedAerosolActivationParameters
-
-Returns a calibrated set of aerosol activation parameters (src in EmulatorModelsExt)
-"""
-function CalibratedAerosolActivationParameters end
 
 end # module AerosolActivation.jl
