+++ test/runtests.jl
@@ -2,6 +2,7 @@ using ClimaParams
 
 include("performance_tests.jl")
 include("aerosol_activation_tests.jl")
+include("aerosol_activation_calibration.jl")
 include("heterogeneous_ice_nucleation_tests.jl")
 include("homogeneous_ice_nucleation_tests.jl")
 include("microphysics0M_tests.jl")
